fridge/requirements.txt
asgiref==3.5.2
Django==3.2.16
django-crispy-forms==1.14.0
django-widget-tweaks==1.4.12
Pillow==9.2.0
pytz==2022.4
sqlparse==0.4.3
typing_extensions==4.4.0
